+++ src/firebench/tools/__init__.py
@@ -50,4 +50,5 @@ from .raster_to_perimeters import array_to_geopolygons
 from .standard_file_utils import (
     current_datetime_iso8601,
     datetime_to_iso8601,
+    read_quantity_from_fb_dataset,
 )